refactor(course): split components into feature subfolders
