+++ frontend/src/app/dashboard/catalog/[id]/page.tsx
@@ -1,4 +1,4 @@
-import CourseDetails from "@/src/features/course/components/CourseDetails";
+import CourseDetails from "@/src/features/course/components/course/CourseDetails";
 
 export default async function CatalogCoursePage({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
