+++ frontend/src/app/dashboard/courses/[id]/page.tsx
@@ -1,4 +1,4 @@
-import CourseDetails from "@/src/features/course/components/CourseDetails";
+import CourseDetails from "@/src/features/course/components/course/CourseDetails";
 
 export default async function MyCoursePage({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
