+++ frontend/src/app/dashboard/instructor/page.tsx
@@ -2,7 +2,7 @@ import { Alert, Button, Group, Stack, Text, Title } from "@mantine/core";
 import { IconPlus } from "@tabler/icons-react";
 import Link from "next/link";
 import { fetchInstructorCourses } from "@/src/features/course/actions/courses";
-import { CourseGrid } from "@/src/features/course/components/CourseGrid";
+import { CourseGrid } from "@/src/features/course/components/course/CourseGrid";
 
 export default async function InstructorDashboard(props: {
   searchParams: Promise<{ page?: string }>;
