+++ frontend/src/features/course/components/challenges/ChallengeCard.module.css
