+++ frontend/src/features/course/components/challenges/ChallengeCard.tsx
