+++ frontend/src/features/course/components/challenges/ChallengeFormFields.tsx
