+++ frontend/src/features/course/components/challenges/ChallengePickerModal.tsx
