+++ frontend/src/features/course/components/challenges/PlayChallengeButton.tsx
