+++ frontend/src/features/course/components/course/CourseCard.module.css
