+++ frontend/src/features/course/components/course/CourseCard.tsx
