+++ frontend/src/features/course/components/course/CourseJourneyCard.tsx
