+++ frontend/src/features/course/components/enrollment/CourseEnrollmentButton.tsx
