+++ frontend/src/features/course/components/enrollment/JoinCourseModal.tsx
