+++ frontend/src/features/course/components/management/CourseChallengeDetailsList.tsx
@@ -14,7 +14,7 @@ import {
   getDifficultyColor,
   getStatusColor,
 } from "@/src/features/course/constants/challengeConstants";
-import PlayChallengeButton from "@/src/features/course/components/PlayChallengeButton";
+import PlayChallengeButton from "@/src/features/course/components/challenges/PlayChallengeButton";
 import { getSanitizedHtml } from "@/src/shared/lib/utils";
 import type { ChallengeDetailResponseDto } from "@/src/features/course/actions/challenges";
