+++ frontend/src/features/course/components/management/InstructorMultiSelect.tsx
