+++ frontend/src/features/course/components/people/CoursePeoplePanel.tsx
